modified schema for latestdevices posts and created a DeviceSchemaHead component for latestdevices
