+++ app/components/LatestDeviceTable.jsx
@@ -9,92 +9,97 @@ import { formatDistanceToNow } from "date-fns";
 import Image from "next/image";
 import { fetchedLatestDevices } from "@/lib/fetchedDevices";
 import Link from "next/link";
+import DeviceSchemaHead from "./DeviceSchemaHead";
+
+const baseURL = process.env.NEXT_PUBLIC_BASE_URL;
 
 const LatestDeviceTable = async ({ device }) => {
   const { posts } = await fetchedLatestDevices();
 
-  const deviceSchema = {
-    "@context": "https://schema.org",
-    "@graph": [
-      {
-        "@type": "BlogPosting",
-        headline: device.title,
-        description,
-        author: {
-          "@type": "Person",
-          name: device.author,
-          url: baseURL,
-        },
-        datePublished: device.publishedAt,
-        image: [metaDataImage],
-        mainEntityOfPage: {
-          "@type": "WebPage",
-          "@id": `${baseURL}/${device.slug}`,
-        },
-        publisher: {
-          "@type": "Organization",
-          name: "Tech Arena24",
-          logo: {
-            "@type": "ImageObject",
-            url: `${baseURL}/images/logoTa24.jpeg`,
-          },
-        },
-        articleSection:
-          device.categories?.map((cat) => cat.title).join(", ") ||
-          "Uncategorized",
-        inLanguage: "en",
-        isAccessibleForFree: true,
-        speakable: {
-          "@type": "SpeakableSpecification",
-          cssSelector: [`${device.tlte}, ${device.summary}`],
-        },
-      },
-      {
-        "@type": "Product",
-        name: device.title,
-        image: [metaDataImage],
-        description,
-        sku: device.sku || device.slug,
-        brand: {
-          "@type": "Brand",
-          name: device.brand || "Generic",
-        },
-        review: {
-          "@type": "Review",
-          author: {
-            "@type": "Person",
-            name: device.author,
-          },
-          datePublished: device.publishedAt,
-          reviewBody: device.summary || "A detailed review of the device.",
-          reviewRating: {
-            "@type": "Rating",
-            ratingValue: device.rating?.toString() || "4.5",
-            bestRating: "5",
-          },
-        },
-        // aggregateRating: {
-        //   "@type": "AggregateRating",
-        //   ratingValue: device.rating?.toString() || "4.5",
-        //   reviewCount: device.reviewCount?.toString() || "24",
-        // },
-        // offers: {
-        //   "@type": "Offer",
-        //   priceCurrency: device.currency || "USD",
-        //   price: device.price || "299",
-        //   itemCondition: "https://schema.org/NewCondition",
-        //   availability: "https://schema.org/InStock",
-        //   url: `${baseURL}/${device.slug}`,
-        // },
-      },
-    ],
-  };
+  // const deviceSchema = {
+  //   "@context": "https://schema.org",
+  //   "@graph": [
+  //     {
+  //       "@type": "BlogPosting",
+  //       headline: device.title,
+  //       description,
+  //       author: {
+  //         "@type": "Person",
+  //         name: device.author,
+  //         url: baseURL,
+  //       },
+  //       datePublished: device.publishedAt,
+  //       image: [metaDataImage],
+  //       mainEntityOfPage: {
+  //         "@type": "WebPage",
+  //         "@id": `${baseURL}/${device.slug}`,
+  //       },
+  //       publisher: {
+  //         "@type": "Organization",
+  //         name: "Tech Arena24",
+  //         logo: {
+  //           "@type": "ImageObject",
+  //           url: `${baseURL}/images/logoTa24.jpeg`,
+  //         },
+  //       },
+  //       articleSection:
+  //         device.categories?.map((cat) => cat.title).join(", ") ||
+  //         "Uncategorized",
+  //       inLanguage: "en",
+  //       isAccessibleForFree: true,
+  //       speakable: {
+  //         "@type": "SpeakableSpecification",
+  //         cssSelector: [`${device.tlte}, ${device.summary}`],
+  //       },
+  //     },
+  //     {
+  //       "@type": "Product",
+  //       name: device.title,
+  //       image: [metaDataImage],
+  //       description,
+  //       sku: device.sku || device.slug,
+  //       brand: {
+  //         "@type": "Brand",
+  //         name: device.brand || "Generic",
+  //       },
+  //       review: {
+  //         "@type": "Review",
+  //         author: {
+  //           "@type": "Person",
+  //           name: device.author,
+  //         },
+  //         datePublished: device.publishedAt,
+  //         reviewBody: device.summary || "A detailed review of the device.",
+  //         reviewRating: {
+  //           "@type": "Rating",
+  //           ratingValue: device.rating?.toString() || "4.5",
+  //           bestRating: "5",
+  //         },
+  //       },
+  //       // aggregateRating: {
+  //       //   "@type": "AggregateRating",
+  //       //   ratingValue: device.rating?.toString() || "4.5",
+  //       //   reviewCount: device.reviewCount?.toString() || "24",
+  //       // },
+  //       // offers: {
+  //       //   "@type": "Offer",
+  //       //   priceCurrency: device.currency || "USD",
+  //       //   price: device.price || "299",
+  //       //   itemCondition: "https://schema.org/NewCondition",
+  //       //   availability: "https://schema.org/InStock",
+  //       //   url: `${baseURL}/${device.slug}`,
+  //       // },
+  //     },
+  //   ],
+  // };
 
   return (
     <>
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(deviceSchema) }}
+      <DeviceSchemaHead
+        device={device}
+        baseURL={baseURL}
+        metaDataImage={posts.deviceImage}
+        description={posts.summary}
       />
       <div className="mb-5">
         <AdBanner />
